@@ -1,66 +1,164 @@
-[tool.poetry]
+[build-system]
+requires = ["hatchling", "hatch-vcs"]
+build-backend = "hatchling.build"
+
+[project]
 name = "jetscape_analysis"
-version = "0.1.0"
-description = "Basic JETSCAPE analysis code"
-license = "BSD-3-Clause"
 authors = [
-    "James Mulligan <james.mulligan@berkeley.edu>",
-    "Raymond Ehlers <raymond.ehlers@cern.ch>",
+  { name = "James Mulligan", email = "james.mulligan@berkeley.edu" },
+  { name = "Raymond Ehlers", email = "raymond.ehlers@cern.ch" },
 ]
+description = "Basic JETSCAPE analysis code"
+readme = "README.md"
+license.file = "LICENSE"
+requires-python = ">=3.9"
 classifiers = [
-    # How mature is this project? Common values are
-    #   3 - Alpha
-    #   4 - Beta
-    #   5 - Production/Stable
-    'Development Status :: 4 - Beta',
-    # Indicate who your project is intended for
-    'Intended Audience :: Science/Research',
-    # Specify the Python versions you support here. In particular, ensure
-    # that you indicate whether you support Python 2, Python 3 or both.
-    'Programming Language :: Python :: 3.6',
-    'Programming Language :: Python :: 3.7',
+  "Development Status :: 4 - Beta",
+  "Intended Audience :: Science/Research",
+  "Intended Audience :: Developers",
+  "License :: OSI Approved :: BSD License",
+  "Operating System :: OS Independent",
+  "Programming Language :: Python",
+  "Programming Language :: Python :: 3",
+  "Programming Language :: Python :: 3 :: Only",
+  "Programming Language :: Python :: 3.9",
+  "Programming Language :: Python :: 3.10",
+  "Programming Language :: Python :: 3.11",
+  "Programming Language :: Python :: 3.12",
+  "Topic :: Scientific/Engineering",
+  "Typing :: Typed",
+]
+dynamic = ["version"]
+dependencies = [
+    "pyyaml >=5.1.2",
+    "numba >=0.59",
+    "h5py >=3.6.0",
+    "tqdm >=4.37.0",
+    "attrs >=20.3.0",
+    "numpy >=1.17.3",
+    "pandas >=2.0",
+    "pyhepmc >=0.4.3",
+    "pyarrow >=3.0.0",
+    "awkward >=1.7.0",
+    "uproot >=5.3",
+    "matplotlib >=3.8",
 ]
 
-[tool.poetry.dependencies]
-python = ">=3.7.1,<3.10"
-tqdm = "^4.37.0"
-numpy = "^1.17.3"
-pyhepmc_ng = "^0.4.3"
-pyyaml = "^5.1.2"
-isort = {extras = ["pyproject"], version = "^4.3.21"}
-awkward = ">=1.7.0"
-pyarrow = "^3.0.0"
-pandas = "^1.2.4"
-attrs = "^20.3.0"
-numba = "^0.53.1"
-h5py = "^3.6.0"
+[project.optional-dependencies]
+dev = [
+    "pytest >=6",
+    "pytest-cov >=3",
+    "pytest-sugar >=1.0.0",
+    "pre-commit >=2.13.0",
+    "ruff >=0.0.209",
+    "mypy >=0.931",
+]
 
-[tool.poetry.dev-dependencies]
-pytest = "^5.2.2"
-flake8 = "^3.7.9"
-flake8-colors = "^0.1.6"
-black = "^19.10b0"
-pre-commit = "^1.20.0"
-mypy = "^0.812"
+[project.urls]
+Homepage = "https://github.com/jdmulligan/JETSCAPE-analysis"
+"Bug Tracker" = "https://github.com/jdmulligan/JETSCAPE-analysis/issues"
+Discussions = "https://github.com/jdmulligan/JETSCAPE-analysis/discussions"
+Changelog = "https://github.com/jdmulligan/JETSCAPE-analysis/releases"
 
-[tool.isort]
-known_standard_library = ["setuptools"]
-known_third_party = ["ROOT", "common_base", "event", "fastjet", "fjext", "pyhepmc_ng", "reader", "tqdm", "yaml"]
-known_first_party = ["heppy"]
-line_length = 120
-# Meaning a grouped grid of imports
-multi_line_output = 3
-lines_after_imports = 1
-force_grid_wrap = 0
-use_parentheses = true
-verbose = true
-skip = ["docs/conf.py"]
-skip_glob = ".*"
+[tool.hatch]
+version.source = "vcs"
+build.hooks.vcs.version-file = "jetscape_analysis/_version.py"
 
-[tool.black]
+[tool.mypy]
+files = ["jetscape_analysis", "tests"]
+python_version = "3.10"
+warn_unused_configs = true
+strict = true
+enable_error_code = ["ignore-without-code", "redundant-expr", "truthy-bool"]
+warn_unreachable = true
+disallow_untyped_defs = false
+disallow_incomplete_defs = false
+
+[[tool.mypy.overrides]]
+module = "jetscape_analysis.*"
+disallow_untyped_defs = true
+disallow_incomplete_defs = true
+
+[[tool.mypy.overrides]]
+module = ["ROOT", "numba", "nox", "scipy.*"]
+ignore_missing_imports = true
+
+[tool.ruff]
+src = ["jetscape_analysis"]
+exclude = [
+    "jetscape_analysis/_version.py",  # Automatically generated, so there's no point in linting it
+    '.git',
+    '.env',
+    '.env3',
+    '.venv',
+    '.venv*',
+    'env',
+    '__pycache__',
+    'bak',
+    'build',
+    'dist',
+]
 line-length = 120
 
-[build-system]
-requires = ["poetry_core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
+[tool.ruff.lint]
+extend-select = [
+  "B",        # flake8-bugbear
+  "I",        # isort
+  "C901",     # mccabe
+  "ARG",      # flake8-unused-arguments
+  "C4",       # flake8-comprehensions
+  "EM",       # flake8-errmsg
+  "ICN",      # flake8-import-conventions
+  "G",        # flake8-logging-format
+  "PGH",      # pygrep-hooks
+  "PIE",      # flake8-pie
+  "PL",       # pylint
+  "PT",       # flake8-pytest-style
+  "PTH",      # flake8-use-pathlib
+  "RET",      # flake8-return
+  "RUF",      # Ruff-specific
+  "SIM",      # flake8-simplify
+  "T20",      # flake8-print
+  "UP",       # pyupgrade
+  "YTT",      # flake8-2020
+  "EXE",      # flake8-executable
+  "NPY",      # NumPy specific rules
+]
+ignore = [
+  "PLR09",    # Too many <...>
+  "PLR2004",  # Magic value used in comparison
+  "ISC001",   # Conflicts with formatter
+  "G004",     # Logging format string. This isn't best practices, but I use it too much to change for now
+  "PLC0414",  # Messes with the re-export of imports needed from mypy
+]
+unfixable = [
+  "T20",      # Removes print statements
+  "F841",     # Removes unused variables
+]
 
+[tool.ruff.lint.mccabe]
+max-complexity = 12
+
+[tool.ruff.lint.isort]
+#required-imports = ["from __future__ import annotations"]
+known-third-party = ["ROOT", "common_base", "event", "fastjet", "fjext", "pyhepmc_ng", "reader", "tqdm", "yaml"]
+known-first-party = ["heppy"]
+
+[tool.ruff.lint.per-file-ignores]
+"tests/**" = ["T20"]
+"noxfile.py" = ["T20"]
+
+
+[tool.pylint]
+py-version = "3.10"
+ignore-paths = [".*/_version.py"]
+reports.output-format = "colorized"
+similarities.ignore-imports = "yes"
+messages_control.disable = [
+  "design",
+  "fixme",
+  "line-too-long",
+  "missing-module-docstring",
+  "wrong-import-position",
+  "logging-fstring-interpolation",  # See above. I use it too much
+]
